pyproject.toml
[project]
name = "mailslide"
version = "0.3.7"
description = "Extract, analyze, and automate Outlook Classic emails on Windows, with optional LLM-powered plugins."
readme = "README.md"
requires-python = ">=3.13"
license = "GPL-3.0-or-later"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.27.0",
    "loguru>=0.7.0",
    "openpyxl>=3.1.5",
    "packaging>=24.2",
    "pycron>=0.2.2",
    "platformdirs>=4.3.8",
    "pywin32>=311; sys_platform == 'win32'",
    "pyyaml>=6.0.3",
    "textual>=8.1.1",
]

[project.scripts]
mailslide = "outlook_mail_extractor.__main__:main"
mailslide-tui = "outlook_mail_extractor.tui:main"

[tool.setuptools.packages.find]
include = ["outlook_mail_extractor*", "mailslide*"]
exclude = ["config*", "tasks*", "tests*"]

[dependency-groups]
dev = [
    "babel>=2.17.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
    "textual-dev>=1.8.0",
    "types-pywin32>=311.0.0.20260317",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.setuptools.package-data]
outlook_mail_extractor = [
    "locales/*.yaml",
    "locales/gettext/**/*.mo",
    "locales/gettext/**/*.po",
    "resources/config_samples/*.yaml.sample",
    "resources/config_samples/plugins/*.yaml.sample",
    "resources/docs/*.md",
]

[tool.mypy]
python_version = "3.13"

[[tool.mypy.overrides]]
module = "pycron"
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
